Fixed IndexError bug. Changed plot interface.
The issue was within the SpecScan.set_x_axis_indices method. The indices
which were being used to access elements of the command list were
incorrect.

The keyword arguments for the SpecScan.doplot() method were moved to the
SpecScan.plot() method, so that the interface could be accessed by the
end user.

In the future, it may be worth investigating whether the SpecPlot class
could be made a subclass of matplotlib.pyplot. Is pyplot even a class?