Implemented OO layer of pyplot. Added new tests.
Implemented the object-oriented layer of pyplot so that title, axis, and
colorbar labels could be edited after the fact. This will allow greater
control over the plot objects which hawpy generates.

Added new test scripts at the bottom of hawpy.py to reflect this change,
and to give the user some examples of how to use the module.