Changed what the SpecPlot object represents.
Rather than representing the data from only one specific scan, the
SpecPlot object will instead represent one Matplotlib figure window.
This way, the data from multiple scans can be plotted in the same
figure.

Added getter and setter methods for the x-axis, y-axis, title, and
colorbar labels. Note that there is not getter method for the colorbar,
as I don't think it is possible to get.